+++ src/common/resolver.ts
